feat: kubernetes deploy and real-time status update (#68)
* dev: generic methods for calculating percentiles

* dev: javadocs style commenting

* dev: change MTTFB percentile map value with Long

* dev: stop status append

* dev: already running errorcode append

* dev: validation update

* dev: agent stop event, avg cal, preftest update

* dev: agent deploy as statefulSet

* dev: status logging

* dev: adjust test method

* dev: stomp socket handling due to k8s

* dev: remove unused ingress

* dev: setup stomp with SockJS

* dev: kubernetes profile agent loader

* dev: docker buildx multiplatform image

* dev: setup EUREKA k8s config

* dev: local image pusher script

* chore: undo issuer

* dev: agent renewal parallel execution

* chore: gradle clean and build

* dev: websocket service exporter

* dev: imagePulling always & eureka connect

* dev: SockJS script src added

* dev: ddl-auto validate to update

* dev: spring cloud kubernetes import

* dev: RBAC role & rolebinding to serviceaccount

* dev: ingress rule update

* dev: limiting resource

* dev: tls issuer and connect to ingress

+++ bm-common/src/main/java/org/benchmarker/bmcommon/dto/CommonTestResult.java
@@ -58,8 +58,7 @@ public class CommonTestResult {
     private String mttfbAverage;
 
     @JsonProperty("mttfb_percentiles")
-    private Map<Double, Double> MTTFBPercentiles;
-
+    private Map<Double, Long> MTTFBPercentiles;
     @JsonProperty("tps_average")
     private double tpsAverage;
